app/requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
ccxt>=4.0.0
PyYAML>=6.0.1
structlog>=24.1.0
python-json-logger>=2.0.7

# Technical Analysis
TA-Lib>=0.4.28
tulipy>=0.4.0
stockstats>=0.5.4

# Notifications
python-telegram-bot>=20.7
twilio>=8.10.0
slackweb>=1.0.5
discord-webhook>=1.0.0
requests>=2.31.0

# Utils
Cython>=3.0.0
jinja2>=3.1.0
tabulate>=0.9.0
tenacity>=8.2.0

# Optional dependencies
python-dotenv>=1.0.0  # для работы с переменными окружения
pathlib>=1.0.1  # для работы с путями
importlib-metadata>=7.0.0  # для работы с метаданными пакетов
